+++ CHANGELOG.md
@@ -4,7 +4,7 @@ seleniumtid changelog
 v0.8.1
 ------
 
-*Release date: 2015-01-??*
+*Release date: 2015-01-26*
 
   * Fixed minor bugs
   * Add visual testing to lettuce tests